+++ src/arcanoidgame/VVector2D.java
@@ -6,23 +6,23 @@ import java.util.Vector;
 
 public class VVector2D {
 
-	private float [] v = new float[2];
+	private double [] v = new double[2];
 
 	public VVector2D() {
 		v[0] = v[1] = 0;
 	}
 
-	public VVector2D( float x, float y ) {
+	public VVector2D( double x, double y ) {
 		v[0] = x;
 		v[1] = y;
 	}
 
 	public VVector2D( Point2D P ) {
-		v[0] = (float)P.getX();
-		v[1] = (float)P.getY();
+		v[0] = P.getX();
+		v[1] = P.getY();
 	}
 
-	public void copy( float x, float y ) {
+	public void copy( double x, double y ) {
 		v[0] = x;
 		v[1] = y;
 	}
@@ -32,36 +32,36 @@ public class VVector2D {
 		v[1] = V.v[1];
 	}
 
-	public float x() { return v[0]; }
-	public float y() { return v[1]; }
+	public double x() { return v[0]; }
+	public double y() { return v[1]; }
 
-	public float lengthSquared() {
+	public double lengthSquared() {
 		return x()*x() + y()*y();
 	}
-	public float length() {
-		return (float)Math.sqrt( lengthSquared() );
+	public double length() {
+		return Math.sqrt( lengthSquared() );
 	}
 
 	// The returned angle is in [0, 2*pi]
-	public float angle() {
+	public double angle() {
 		return angle( 0 );
 	}
 
 	// The returned angle is in [lowerBound, lowerBound+2*pi]
-	public float angle(
-		float lowerBound // in radians
+	public double angle(
+		double lowerBound // in radians
 	) {
-		float l = length();
+		double l = length();
 		if ( l <= 0 )
 			return lowerBound;
-		float angle = (float)Math.asin(y()/l);
+		double angle = Math.asin(y()/l);
 		// Now angle is in [-pi/2,pi/2]
 		if ( x() < 0 )
-			angle = (float)Math.PI - angle;
+			angle = Math.PI - angle;
 		// Now angle is in [-pi/2,3*pi/2]
 
-		float TwoPi = (float)( 2*Math.PI );
-		float upperBound = lowerBound + TwoPi;
+		double TwoPi = ( 2*Math.PI );
+		double upperBound = lowerBound + TwoPi;
 		while ( angle > upperBound )
 			angle -= TwoPi;
 		while ( angle < lowerBound )
@@ -75,16 +75,16 @@ public class VVector2D {
 	}
 
 	public VVector2D normalized() {
-		float l = length();
+		double l = length();
 		if ( l > 0 ) {
-			float k = 1/l; // scale factor
+			double k = 1/l; // scale factor
 			return new VVector2D(k*x(),k*y());
 		}
 		else return new VVector2D(x(),y());
 	}
 
 	// returns the dot-product of the given vectors
-	static public float dot( VVector2D a, VVector2D b ) {
+	static public double dot( VVector2D a, VVector2D b ) {
 		return a.x()*b.x() + a.y()*b.y();
 	}
 
@@ -107,14 +107,14 @@ public class VVector2D {
 	// The angle returned is in the interval [-pi,pi],
 	// where a positive angle corresponds to a counterclockwise rotation
 	// (assuming x+ right, y+ up).
-	static public float computeSignedAngle( VVector2D v1, VVector2D v2 ) {
+	static public double computeSignedAngle( VVector2D v1, VVector2D v2 ) {
 		//return Vector3D.computeSignedAngle(
 		//	new Vector3D( v1.x(), v1.y(), 0 ),
 		//	new Vector3D( v2.x(), v2.y(), 0 ),
 		//	new Vector3D( 0, 0, 1 )
 		//);
 
-		float productOfLengths = v1.length() * v2.length();
+		double productOfLengths = v1.length() * v2.length();
 		if ( productOfLengths <= 0 )
 			return 0;
 
@@ -131,7 +131,7 @@ public class VVector2D {
 		double angle = ( sineOfAngle >= 1 ) ? Math.PI/2 : Math.asin( sineOfAngle );
 
 		// Compute the dot product of v1 and v2
-		float dotProduct = VVector2D.dot( v1, v2 );
+		double dotProduct = VVector2D.dot( v1, v2 );
 
 		if ( dotProduct < 0 )
 			angle = Math.PI - angle;
@@ -139,11 +139,11 @@ public class VVector2D {
 		if ( crossProduct_z < 0 )
 			angle = - angle;
 
-		return (float)angle;
+		return angle;
 	}
 
         public static VVector2D reflect(VVector2D vector, VVector2D normal){
-            float vDotX2 = VVector2D.dot(vector,normal)*2;
-            return vector.diff(vector,vector.mult(normal,vDotX2)); 
+            double vDotX2 = VVector2D.dot(vector,normal)*2;
+            return vector.diff(vector,vector.mult(normal, (float) vDotX2)); 
         }
 }
